Restored prev FileProgress location and use with
